refactor(tests): replace projectStore.test with projectStore.test.tsx and enhance tests
